ConvertToMesh for text ( and possibly curves ) was generating the wrong normals - the fact that the generated model was being set to double sided helped hide this.
To test for this in Blender, just add text, TAB out of edit mode, then convert to mesh ( ALT+C ), and press P for the game engine.  Alternatively select Textured Draw Mode.  You'll notice that the text is only visible from the back.  This patch reverses this, to be more correct.

Based on discussions on IRC, I'm now trying to fix another aspect of the extrude code - the fact that the faces at the front and back of the extruded curve face the same way ( ie one of them will be wrong ).  I'll keep working on this, but if someone can help out, feel free!

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@12349 954f8c5b-7b00-dc11-b283-0030488c597c