Mesh Deform Modifier ====================
Dynamic binding support. This means that the mesh can move _within_
the cage and still deform correct. If the mesh goes out of the cage,
don't expect correct result. Must be enabled with the 'Dynamic'
option, because it is slower and consumes more memory.

This is useful to use e.g. the cage mesh for main deformations and
still have shape keys for facial deformation working.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@12559 954f8c5b-7b00-dc11-b283-0030488c597c