* New feature on compo scale node: "Scene Size %"
This option sets the relative scaling factor to the amount set in the 
scene "100%/75%/50%/25%" buttons. It's useful when you've got a fixed 
background image, and want to do preview renders at a lesser 
percentage, so you don't have to go and change the scale node each 
time you change the %.

Also removed unnecessary use of a global from texture node.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@12583 954f8c5b-7b00-dc11-b283-0030488c597c