Merge from Harmonic Skeleton branch
This code adds a basic and simple skeleton generator.

Examples and links are in the wiki, docs will come eventually: http://wiki.blender.org/index.php/User:Theeth/skeletor

In a nutshell, select a vertex at the top of the head and press "Generate Skeleton". UI Panel is in the Editing buttons in Edit Mode, tooltips and semi-useful.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@12838 954f8c5b-7b00-dc11-b283-0030488c597c