== Action Editor - Sample Keys ==
Cessen Request - Added a tool which inserts keyframes on every frame between two selected keyframes. This is useful for adding overlapping motion (apparently). 

Use Alt-O to run this tool. The name of this tool is also open to suggestions ;)

Note:
There is currently a bug with this. When there are 4 consecutive keyframes, it won't add keyframes between keyframes 2 and 3.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@13169 954f8c5b-7b00-dc11-b283-0030488c597c