=== Transform Snap ===
Snapping Mode: Active

With this mode, the active element (at this time, object or vertice) is used as snapping target. If there is no active element in the selection, it reverts back to median mode.


Edit Mode snapping, other meshes no longer have to be selected to act as snapping point.


Fix a potential bug with snapping point from other meshes.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@13318 954f8c5b-7b00-dc11-b283-0030488c597c