== Limit Rotation Constraint - Affects Transform ==
Now, the Limit Rotation Constraint can be made to work on the Transform values like the Limit Location/Rotation constraints. Use the "For Transform" button to activate.

Note: this does not really work as well as for the other constraints, as it currently does conversions between eulers/quats and matrices. This results in loss of rotation data.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@13325 954f8c5b-7b00-dc11-b283-0030488c597c