== Mirror Modifier & Subdivide ==
Vertices that are newly created by a subdivision of a mesh object
with mirror modifiers (with clipping option set) are projected to
the plane of symmetry when they are created on edges that lie on
that plane.

For regular subdivide this is trivial by linearity, so this mainly
effects subdivide smooth, loopcut smooth, and subdivide fractal.
This commit basically prevents nasty seam rips on the mirror when
doing these operations.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@13407 954f8c5b-7b00-dc11-b283-0030488c597c