Fix for indirectly linked libraries.
When a.blend -> //../c.blend -> //d.blend
Blender would see d.blend as //d.blend which is wrong since that is relative to c.blend
This works, except when you make an indirectly linked group into a directly linked group. and then the incorrect path was saved into the blend file and the data not load up again (without copying it there or hex editing the path)

This loop makes all lib files relative to the blend you have open.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@13505 954f8c5b-7b00-dc11-b283-0030488c597c