== PyNodes ==
First fix in trunk: The "Update" button that appears in the Material buttons -> Nodes panel when a pynode is selected wasn't forcing an immediate update. Added a BIF_preview_changed call to fix it.

Let's pretend I actually knew that button also appeared there :) ... weeks working on pynodes and only right after Nathan commits to trunk I realize that this part of node panels is drawn in the buttons window, too.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@13621 954f8c5b-7b00-dc11-b283-0030488c597c