== Align to Transform Orientation ==
New rotation alignement fonction

Rotates objects/Pose bones to match the selected transform orientation.

Can be used to align to view, active object (normal) and custom transform orientations.

Accessible in the Object -> Transform submenu and through the hotkey Ctrl-Alt-A (which was previously a fall through for Apply but only Ctrl-A and Ctrl-Shift-A did anything special).

Can be eventually made to work in edit mode (not too hard).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@14037 954f8c5b-7b00-dc11-b283-0030488c597c