Insert Keyframe Optimisations:
Now a binary search is performed instead of a linear one to see where to insert a keyframe. It also checks first whether the keyframe is out of the bounds of the existing ones (as most of the time, keyframes are inserted at the end of the array). 

When using the .BVH importer to import a particularly large file, the time taken to add the keyframes improved by about 1 second. Other factors probably limited the improvement seen.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@14068 954f8c5b-7b00-dc11-b283-0030488c597c