[#7289] Orbit around selected in edit mode goes crazy with no verts selected [#8784] orbiting around selection bug + a few kind of a must features (partial fix)
As suggested by Ton, keep last center and use that when nothing selected.

I don't really like it though as it wouldn't work ok when changing scene  or when using a different 3D view and is somewhat non trivial to explain (it's not the last selection, it's the last used orbiting center).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@14281 954f8c5b-7b00-dc11-b283-0030488c597c