== Armature Tools - Separate (not yet functional) ==
Initial commit of Separate tool for Armatures. Currently, the functionality is hidden behind a temporary patch, as there are still issues to be worked out (crashes under certain conditions and a re-linking issue). It may remain like this for the release if I can't get it to work correctly.

Note:
- Hotkey for separate is Ctrl-Shift-P (it's a bit clumsy, and isn't consistent with P for separate for meshes, but Select Parent(s) is better as P)

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@14290 954f8c5b-7b00-dc11-b283-0030488c597c