== RED one (redcode) ==
This adds redcode (the file format of RED one, R3D) support to blender.
Seems to work fine with the footage I found on the web, but keep in
mind, that because of the unoptimized nature of libopenjpeg, frame
decoding isn't that fast.

It is also a rather challenging task, to make 4k-float-footage realtime :)



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@15089 954f8c5b-7b00-dc11-b283-0030488c597c