Fix for GE Make Runtime menu option
Doh - I forgot to add a comment to the previous commit.

This adds a file exists check for Make Dynamic Runtime for the GE.  It also removes this menu option for now, as the required file ( blenderdynplayer.exe ) hasn't been included with the Blender app for quite a few releases ( therefore the feature has been broken ).

I also renamed the option, to make it a bit more understandable.  This might even be better moved to the Game menu, but I didn't want to do that ( yet :) )

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@15910 954f8c5b-7b00-dc11-b283-0030488c597c