== Grease Pencil ==
Special request by Alxarch for Architecture:
Hold Ctrl-Key when 'Draw Mode' is enabled to draw straight lines. 
Although when drawing the stroke, the stroke will still be freehand, the final result will be a line between the endpoints of that stroke. This is useful for annotations of sectioning lines + site maps, etc.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@16298 954f8c5b-7b00-dc11-b283-0030488c597c